# src/main/resources/openapi/products.yaml
openapi: 3.0.3
info:
  title: Amancay Products API
  version: 0.0.1
paths:
  /api/products:
    get:
      summary: List products
      parameters:
        - {name: page, in: query, schema: {type: integer, default: 0}}
        - {name: size, in: query, schema: {type: integer, default: 20}}
        - {name: q, in: query, schema: {type: string}}
        - {name: is_active, in: query, schema: {type: boolean}}
      responses:
        '200': {description: Product page}
    post:
      summary: Create a product with variants
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateProductRequest'}
      responses:
        '201': {description: Product created}
  /api/products/{id}:
    get:
      summary: Get a product with variants
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        '200': {description: Product}
    put:
      summary: Update a product and synchronize variants
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        '200': {description: Product updated}
    delete:
      summary: Delete a product and its variants
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        '204': {description: Product deleted}
components:
  schemas:
    ProductVariant:
      type: object
      properties:
        id: {type: string, format: uuid}
        price: {type: number, format: decimal, minimum: 0}
        stockQuantity: {type: integer, minimum: 0}
    Product:
      allOf:
        - $ref: '#/components/schemas/ProductSummary'
        - type: object
          properties:
            shortDescription: {type: string}
            description: {type: string}
            variants: {type: array, items: {$ref: '#/components/schemas/ProductVariant'}}
    ProductSummary:
      type: object
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        slug: {type: string}
        active: {type: boolean}
    CreateProductRequest:
      type: object
      required: [name, slug]
      properties:
        name: {type: string, maxLength: 255}
        slug: {type: string, maxLength: 255}
        shortDescription: {type: string}
        description: {type: string}
        active: {type: boolean}
        variants: {type: array, items: {$ref: '#/components/schemas/ProductVariant'}}